+++ raster_analysis/geometry.py
@@ -1,10 +1,12 @@
+import sys
 from typing import Any, Dict
 
-from shapely.geometry import shape, Polygon
+import geobuf
+from shapely.geometry import shape, Polygon, mapping
 
 from raster_analysis.exceptions import InvalidGeometryException
 from raster_analysis.utils import decode_geometry
-from raster_analysis.globals import BasePolygon
+from raster_analysis.globals import BasePolygon, LAMBDA_ASYNC_PAYLOAD_LIMIT_BYTES
 
 
 class GeometryTile:
@@ -35,4 +37,31 @@ class GeometryTile:
                 self.geom = {}
 
             self.geom: BasePolygon = geom_tile
+
+
+def encode_geometry(geom: BasePolygon) -> str:
+    """
+    Encode geometry into a compressed string
+    """
+    encoded_geom = geobuf.encode(mapping(geom)).hex()
+
+    # if the geometry is so complex is still goes over the limit, incrementally attempting to simplify it
+    if sys.getsizeof(encoded_geom) > LAMBDA_ASYNC_PAYLOAD_LIMIT_BYTES:
+        encoded_geom = geobuf.encode(
+            mapping(geom.simplify(0.005, preserve_topology=False))
+        ).hex()
+
+    if sys.getsizeof(encoded_geom) > LAMBDA_ASYNC_PAYLOAD_LIMIT_BYTES:
+        encoded_geom = geobuf.encode(
+            mapping(geom.simplify(0.01, preserve_topology=False))
+        ).hex()
+
+    return encoded_geom
+
+
+def decode_geometry(geom: str) -> BasePolygon:
+    """
+    Decode geometry from compressed string
+    """
+    return shape(geobuf.decode(bytes.fromhex(geom))).buffer(0)
